requirements.txt
shiny
pandas
matplotlib
ipyleaflet
shinywidgets
googlesearch-python
geopandas